Decouple DNS.
