+++ include/FreeRTOS_IP_Private.h
@@ -19,31 +19,39 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  *
- * http://aws.amazon.com/freertos
- * http://www.FreeRTOS.org
+ * https://github.com/FreeRTOS
+ * https://www.FreeRTOS.org
  */
 
 #ifndef FREERTOS_IP_PRIVATE_H
-    #define FREERTOS_IP_PRIVATE_H
+#define FREERTOS_IP_PRIVATE_H
 
-    #ifdef __cplusplus
-        extern "C" {
-    #endif
+/* *INDENT-OFF* */
+#ifdef __cplusplus
+    extern "C" {
+#endif
+/* *INDENT-ON* */
 
 /* Application level configuration options. */
-    #include "FreeRTOSIPConfig.h"
-    #include "FreeRTOSIPConfigDefaults.h"
-    #include "FreeRTOS_Sockets.h"
-    #include "IPTraceMacroDefaults.h"
-    #include "FreeRTOS_Stream_Buffer.h"
-    #if ( ipconfigUSE_TCP == 1 )
-        #include "FreeRTOS_TCP_WIN.h"
-        #include "FreeRTOS_TCP_IP.h"
-    #endif
+#include "FreeRTOSIPConfig.h"
+#include "FreeRTOSIPConfigDefaults.h"
+#include "FreeRTOS_Sockets.h"
+#include "IPTraceMacroDefaults.h"
+#include "FreeRTOS_Stream_Buffer.h"
+#if ( ipconfigUSE_TCP == 1 )
+    #include "FreeRTOS_TCP_WIN.h"
+    #include "FreeRTOS_TCP_IP.h"
+#endif
 
-    #include "semphr.h"
+#include "semphr.h"
 
-    #include "event_groups.h"
+#include "event_groups.h"
+
+#ifdef TEST
+    int ipFOREVER( void );
+#else
+    #define ipFOREVER()    1
+#endif
 
 /*-----------------------------------------------------------*/
 /* Utility macros for marking casts as recognized during     */
@@ -51,535 +59,530 @@
 /* Note _HT_ Changed 'vCastConstPointerTo' to the shorter    */
 /* vCastConstPtrTo to limit the length of the function name. */
 /*-----------------------------------------------------------*/
-    #define ipCAST_PTR_TO_TYPE_PTR( TYPE, pointer )                ( vCastPointerTo_ ## TYPE( ( void * ) ( pointer ) ) )
-    #define ipCAST_CONST_PTR_TO_CONST_TYPE_PTR( TYPE, pointer )    ( vCastConstPtrTo_ ## TYPE( ( const void * ) ( pointer ) ) )
+#define ipCAST_PTR_TO_TYPE_PTR( TYPE, pointer )                ( vCastPointerTo_ ## TYPE( ( void * ) ( pointer ) ) )
+#define ipCAST_CONST_PTR_TO_CONST_TYPE_PTR( TYPE, pointer )    ( vCastConstPtrTo_ ## TYPE( ( const void * ) ( pointer ) ) )
 
 /*-----------------------------------------------------------*/
 /* Utility macros for declaring cast utility functions in    */
 /* order to centralize typecasting for static analysis.      */
 /*-----------------------------------------------------------*/
-    #define ipDECL_CAST_PTR_FUNC_FOR_TYPE( TYPE )          TYPE * vCastPointerTo_ ## TYPE( void * pvArgument )
-    #define ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( TYPE )    const TYPE * vCastConstPtrTo_ ## TYPE( const void * pvArgument )
+#define ipDECL_CAST_PTR_FUNC_FOR_TYPE( TYPE )          TYPE * vCastPointerTo_ ## TYPE( void * pvArgument )
+#define ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( TYPE )    const TYPE * vCastConstPtrTo_ ## TYPE( const void * pvArgument )
 
-    extern BaseType_t xTCPWindowLoggingLevel;
-    extern QueueHandle_t xNetworkEventQueue;
+extern BaseType_t xTCPWindowLoggingLevel;
+extern QueueHandle_t xNetworkEventQueue;
 
 /*-----------------------------------------------------------*/
 /* Protocol headers.                                         */
 /*-----------------------------------------------------------*/
 
+#include "pack_struct_start.h"
+struct xETH_HEADER
+{
+    MACAddress_t xDestinationAddress;     /**< Destination address  0 + 6 = 6  */
+    MACAddress_t xSourceAddress;          /**< Source address       6 + 6 = 12 */
+    uint16_t usFrameType;                 /**< The EtherType field 12 + 2 = 14 */
+}
+#include "pack_struct_end.h"
+typedef struct xETH_HEADER EthernetHeader_t;
+
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( EthernetHeader_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( EthernetHeader_t );
+
+#include "pack_struct_start.h"
+struct xARP_HEADER
+{
+    uint16_t usHardwareType;                  /**< Network Link Protocol type                     0 +  2 =  2 */
+    uint16_t usProtocolType;                  /**< The internetwork protocol                      2 +  2 =  4 */
+    uint8_t ucHardwareAddressLength;          /**< Length in octets of a hardware address         4 +  1 =  5 */
+    uint8_t ucProtocolAddressLength;          /**< Length in octets of the internetwork protocol  5 +  1 =  6 */
+    uint16_t usOperation;                     /**< Operation that the sender is performing        6 +  2 =  8 */
+    MACAddress_t xSenderHardwareAddress;      /**< Media address of the sender                    8 +  6 = 14 */
+    uint8_t ucSenderProtocolAddress[ 4 ];     /**< Internetwork address of sender                14 +  4 = 18  */
+    MACAddress_t xTargetHardwareAddress;      /**< Media address of the intended receiver        18 +  6 = 24  */
+    uint32_t ulTargetProtocolAddress;         /**< Internetwork address of the intended receiver 24 +  4 = 28  */
+}
+#include "pack_struct_end.h"
+typedef struct xARP_HEADER ARPHeader_t;
+
+#include "pack_struct_start.h"
+struct xIP_HEADER
+{
+    uint8_t ucVersionHeaderLength;            /**< The version field + internet header length 0 + 1 =  1 */
+    uint8_t ucDifferentiatedServicesCode;     /**< Differentiated services code point + ECN    1 + 1 =  2 */
+    uint16_t usLength;                        /**< Entire Packet size, ex. Ethernet header.   2 + 2 =  4 */
+    uint16_t usIdentification;                /**< Identification field                       4 + 2 =  6 */
+    uint16_t usFragmentOffset;                /**< Fragment flags and fragment offset         6 + 2 =  8 */
+    uint8_t ucTimeToLive;                     /**< Time to live field                         8 + 1 =  9 */
+    uint8_t ucProtocol;                       /**< Protocol used in the IP-datagram           9 + 1 = 10 */
+    uint16_t usHeaderChecksum;                /**< Checksum of the IP-header                 10 + 2 = 12 */
+    uint32_t ulSourceIPAddress;               /**< IP address of the source                  12 + 4 = 16 */
+    uint32_t ulDestinationIPAddress;          /**< IP address of the destination             16 + 4 = 20 */
+}
+#include "pack_struct_end.h"
+typedef struct xIP_HEADER IPHeader_t;
+
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPHeader_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPHeader_t );
+
+#if ( ipconfigUSE_IPv6 != 0 )
     #include "pack_struct_start.h"
-    struct xETH_HEADER
+    struct xIP_HEADER_IPv6
     {
-        MACAddress_t xDestinationAddress; /**< Destination address  0 + 6 = 6  */
-        MACAddress_t xSourceAddress;      /**< Source address       6 + 6 = 12 */
-        uint16_t usFrameType;             /**< The EtherType field 12 + 2 = 14 */
+        uint8_t ucVersionTrafficClass;          /**< The version field.                      0 +  1 =  1 */
+        uint8_t ucTrafficClassFlow;             /**< Traffic class and flow.                 1 +  1 =  2 */
+        uint16_t usFlowLabel;                   /**< Flow label.                             2 +  2 =  4 */
+        uint16_t usPayloadLength;               /**< Number of bytes after the IPv6 header.  4 +  2 =  6 */
+        uint8_t ucNextHeader;                   /**< Next header: TCP, UDP, or ICMP.         6 +  1 =  7 */
+        uint8_t ucHopLimit;                     /**< Replaces the time to live from IPv4.    7 +  1 =  8 */
+        IPv6_Address_t xSourceAddress;          /**< The IPv6 address of the sender.         8 + 16 = 24 */
+        IPv6_Address_t xDestinationAddress;     /**< The IPv6 address of the receiver.      24 + 16 = 40 */
     }
     #include "pack_struct_end.h"
-    typedef struct xETH_HEADER EthernetHeader_t;
+    typedef struct xIP_HEADER_IPv6 IPHeader_IPv6_t;
+#endif /* ipconfigUSE_IPv6 */
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( EthernetHeader_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( EthernetHeader_t );
+#if ( ipconfigUSE_IPv6 != 0 )
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPHeader_IPv6_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPHeader_IPv6_t );
+#endif
 
+#include "pack_struct_start.h"
+struct xICMP_HEADER
+{
+    uint8_t ucTypeOfMessage;       /**< The ICMP type                     0 + 1 = 1 */
+    uint8_t ucTypeOfService;       /**< The ICMP subtype                  1 + 1 = 2 */
+    uint16_t usChecksum;           /**< The checksum of whole ICMP packet 2 + 2 = 4 */
+    uint16_t usIdentifier;         /**< Used in some types of ICMP        4 + 2 = 6 */
+    uint16_t usSequenceNumber;     /**< Used in some types of ICMP        6 + 2 = 8 */
+}
+#include "pack_struct_end.h"
+typedef struct xICMP_HEADER ICMPHeader_t;
+
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ICMPHeader_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ICMPHeader_t );
+
+#if ( ipconfigUSE_IPv6 != 0 )
     #include "pack_struct_start.h"
-    struct xARP_HEADER
+    struct xICMPHeader_IPv6
     {
-        uint16_t usHardwareType;              /**< Network Link Protocol type                     0 +  2 =  2 */
-        uint16_t usProtocolType;              /**< The internetwork protocol                      2 +  2 =  4 */
-        uint8_t ucHardwareAddressLength;      /**< Length in octets of a hardware address         4 +  1 =  5 */
-        uint8_t ucProtocolAddressLength;      /**< Length in octets of the internetwork protocol  5 +  1 =  6 */
-        uint16_t usOperation;                 /**< Operation that the sender is performing        6 +  2 =  8 */
-        MACAddress_t xSenderHardwareAddress;  /**< Media address of the sender                    8 +  6 = 14 */
-        uint8_t ucSenderProtocolAddress[ 4 ]; /**< Internetwork address of sender                14 +  4 = 18  */
-        MACAddress_t xTargetHardwareAddress;  /**< Media address of the intended receiver        18 +  6 = 24  */
-        uint32_t ulTargetProtocolAddress;     /**< Internetwork address of the intended receiver 24 +  4 = 28  */
+        uint8_t ucTypeOfMessage;         /**< The message type.     0 +  1 = 1 */
+        uint8_t ucTypeOfService;         /**< Type of service.      1 +  1 = 2 */
+        uint16_t usChecksum;             /**< Checksum.             2 +  2 = 4 */
+        uint32_t ulReserved;             /**< Reserved.             4 +  4 = 8 */
+        IPv6_Address_t xIPv6Address;     /**< The IPv6 address.     8 + 16 = 24 */
+        uint8_t ucOptionType;            /**< The option type.     24 +  1 = 25 */
+        uint8_t ucOptionLength;          /**< The option length.   25 +  1 = 26 */
+        uint8_t ucOptionBytes[ 6 ];      /**< Option bytes.        26 +  6 = 32 */
     }
     #include "pack_struct_end.h"
-    typedef struct xARP_HEADER ARPHeader_t;
+    typedef struct xICMPHeader_IPv6 ICMPHeader_IPv6_t;
+#endif /* ipconfigUSE_IPv6 */
 
+#if ( ipconfigUSE_IPv6 != 0 )
     #include "pack_struct_start.h"
-    struct xIP_HEADER
+    struct xICMPEcho_IPv6
     {
-        uint8_t ucVersionHeaderLength;        /**< The version field + internet header length 0 + 1 =  1 */
-        uint8_t ucDifferentiatedServicesCode; /**< Differentiated services code point + ECN    1 + 1 =  2 */
-        uint16_t usLength;                    /**< Entire Packet size, ex. Ethernet header.   2 + 2 =  4 */
-        uint16_t usIdentification;            /**< Identification field                       4 + 2 =  6 */
-        uint16_t usFragmentOffset;            /**< Fragment flags and fragment offset         6 + 2 =  8 */
-        uint8_t ucTimeToLive;                 /**< Time to live field                         8 + 1 =  9 */
-        uint8_t ucProtocol;                   /**< Protocol used in the IP-datagram           9 + 1 = 10 */
-        uint16_t usHeaderChecksum;            /**< Checksum of the IP-header                 10 + 2 = 12 */
-        uint32_t ulSourceIPAddress;           /**< IP address of the source                  12 + 4 = 16 */
-        uint32_t ulDestinationIPAddress;      /**< IP address of the destination             16 + 4 = 20 */
+        uint8_t ucTypeOfMessage;       /**< The message type.     0 +  1 = 1 */
+        uint8_t ucTypeOfService;       /**< Type of service.      1 +  1 = 2 */
+        uint16_t usChecksum;           /**< Checksum.             2 +  2 = 4 */
+        uint16_t usIdentifier;         /**< Identifier.           4 +  2 = 6 */
+        uint16_t usSequenceNumber;     /**< Sequence number.      6 +  2 = 8 */
     }
     #include "pack_struct_end.h"
-    typedef struct xIP_HEADER IPHeader_t;
+    typedef struct xICMPEcho_IPv6 ICMPEcho_IPv6_t;
+#endif /* ipconfigUSE_IPv6 */
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPHeader_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPHeader_t );
+#include "pack_struct_start.h"
+struct xICMPRouterSolicitation_IPv6
+{
+    uint8_t ucTypeOfMessage;     /**<  0 +  1 =  1 */
+    uint8_t ucTypeOfService;     /**<  1 +  1 =  2 */
+    uint16_t usChecksum;         /**<  2 +  2 =  4 */
+    uint32_t ulReserved;         /**<  4 +  4 =  8 */
+}
+#include "pack_struct_end.h"
+typedef struct xICMPRouterSolicitation_IPv6 ICMPRouterSolicitation_IPv6_t;
 
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #include "pack_struct_start.h"
-        struct xIP_HEADER_IPv6
-        {
-            uint8_t ucVersionTrafficClass;      /**< The version field.                      0 +  1 =  1 */
-            uint8_t ucTrafficClassFlow;         /**< Traffic class and flow.                 1 +  1 =  2 */
-            uint16_t usFlowLabel;               /**< Flow label.                             2 +  2 =  4 */
-            uint16_t usPayloadLength;           /**< Number of bytes after the IPv6 header.  4 +  2 =  6 */
-            uint8_t ucNextHeader;               /**< Next header: TCP, UDP, or ICMP.         6 +  1 =  7 */
-            uint8_t ucHopLimit;                 /**< Replaces the time to live from IPv4.    7 +  1 =  8 */
-            IPv6_Address_t xSourceAddress;      /**< The IPv6 address of the sender.         8 + 16 = 24 */
-            IPv6_Address_t xDestinationAddress; /**< The IPv6 address of the receiver.      24 + 16 = 40 */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xIP_HEADER_IPv6 IPHeader_IPv6_t;
-    #endif /* ipconfigUSE_IPv6 */
+#if ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 )
+    #include "pack_struct_start.h"
+    struct xICMPRouterAdvertisement_IPv6
+    {
+        uint8_t ucTypeOfMessage;           /*  0 +  1 =  1 */
+        uint8_t ucTypeOfService;           /*  1 +  1 =  2 */
+        uint16_t usChecksum;               /*  2 +  2 =  4 */
+        uint8_t ucHopLimit;                /*  4 +  1 =  5 */
+        uint8_t ucFlags;                   /*  5 +  1 =  6 */
+        uint16_t usLifetime;               /*  6 +  2 =  8 */
+        uint16_t usReachableTime[ 2 ];     /*  8 +  4 = 12 */
+        uint16_t usRetransTime[ 2 ];       /* 12 +  4 = 16 */
+    }
+    #include "pack_struct_end.h"
+    typedef struct xICMPRouterAdvertisement_IPv6 ICMPRouterAdvertisement_IPv6_t;
+#endif /* ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 ) */
 
-    #if ( ipconfigUSE_IPv6 != 0 )
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPHeader_IPv6_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPHeader_IPv6_t );
+#if ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 )
+    /* This is an option with the Router Advertisement. */
+    #include "pack_struct_start.h"
+    struct xICMPPrefixOption_IPv6
+    {
+        uint8_t ucType;                   /*  0 +  1 =  1 */
+        uint8_t ucLength;                 /*  1 +  1 =  2 */
+        uint8_t ucPrefixLength;           /*  2 +  1 =  3 */
+        uint8_t ucFlags;                  /*  3 +  1 =  4 */
+        uint32_t ulValidLifeTime;         /*  4 +  4 =  8 */
+        uint32_t ulPreferredLifeTime;     /*  8 +  4 = 12 */
+        uint32_t ulReserved;              /* 12 +  4 = 16 */
+        uint8_t ucPrefix[ 16 ];           /* 16 + 16 = 32 */
+    }
+    #include "pack_struct_end.h"
+    typedef struct xICMPPrefixOption_IPv6 ICMPPrefixOption_IPv6_t;
+#endif /* if ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 ) */
+
+#include "pack_struct_start.h"
+struct xUDP_HEADER
+{
+    uint16_t usSourcePort;          /**< The source port                      0 + 2 = 2 */
+    uint16_t usDestinationPort;     /**< The destination port                 2 + 2 = 4 */
+    uint16_t usLength;              /**< The size of the whole UDP packet     4 + 2 = 6 */
+    uint16_t usChecksum;            /**< The checksum of the whole UDP Packet 6 + 2 = 8 */
+}
+#include "pack_struct_end.h"
+typedef struct xUDP_HEADER UDPHeader_t;
+
+#include "pack_struct_start.h"
+struct xTCP_HEADER
+{
+    uint16_t usSourcePort;                           /**< The Source port                       0 +  2 =  2 */
+    uint16_t usDestinationPort;                      /**< The destination port                  2 +  2 =  4 */
+    uint32_t ulSequenceNumber;                       /**< The Sequence number                   4 +  4 =  8 */
+    uint32_t ulAckNr;                                /**< The acknowledgement number            8 +  4 = 12 */
+    uint8_t ucTCPOffset;                             /**< The value of TCP offset              12 +  1 = 13 */
+    uint8_t ucTCPFlags;                              /**< The TCP-flags field                  13 +  1 = 14 */
+    uint16_t usWindow;                               /**< The size of the receive window       14 +  2 = 15 */
+    uint16_t usChecksum;                             /**< The checksum of the header           15 +  2 = 18 */
+    uint16_t usUrgent;                               /**< Pointer to the last urgent data byte 18 +  2 = 20 */
+    #if ipconfigUSE_TCP == 1
+        uint8_t ucOptdata[ ipSIZE_TCP_OPTIONS ];     /**< The options + 12 = 32 */
     #endif
-
-    #include "pack_struct_start.h"
-    struct xICMP_HEADER
-    {
-        uint8_t ucTypeOfMessage;   /**< The ICMP type                     0 + 1 = 1 */
-        uint8_t ucTypeOfService;   /**< The ICMP subtype                  1 + 1 = 2 */
-        uint16_t usChecksum;       /**< The checksum of whole ICMP packet 2 + 2 = 4 */
-        uint16_t usIdentifier;     /**< Used in some types of ICMP        4 + 2 = 6 */
-        uint16_t usSequenceNumber; /**< Used in some types of ICMP        6 + 2 = 8 */
-    }
-    #include "pack_struct_end.h"
-    typedef struct xICMP_HEADER ICMPHeader_t;
-
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ICMPHeader_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ICMPHeader_t );
-
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #include "pack_struct_start.h"
-        struct xICMPHeader_IPv6
-        {
-            uint8_t ucTypeOfMessage;     /**< The message type.     0 +  1 = 1 */
-            uint8_t ucTypeOfService;     /**< Type of service.      1 +  1 = 2 */
-            uint16_t usChecksum;         /**< Checksum.             2 +  2 = 4 */
-            uint32_t ulReserved;         /**< Reserved.             4 +  4 = 8 */
-            IPv6_Address_t xIPv6Address; /**< The IPv6 address.     8 + 16 = 24 */
-            uint8_t ucOptionType;        /**< The option type.     24 +  1 = 25 */
-            uint8_t ucOptionLength;      /**< The option length.   25 +  1 = 26 */
-            uint8_t ucOptionBytes[ 6 ];  /**< Option bytes.        26 +  6 = 32 */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xICMPHeader_IPv6 ICMPHeader_IPv6_t;
-    #endif /* ipconfigUSE_IPv6 */
-
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #include "pack_struct_start.h"
-        struct xICMPEcho_IPv6
-        {
-            uint8_t ucTypeOfMessage;   /**< The message type.     0 +  1 = 1 */
-            uint8_t ucTypeOfService;   /**< Type of service.      1 +  1 = 2 */
-            uint16_t usChecksum;       /**< Checksum.             2 +  2 = 4 */
-            uint16_t usIdentifier;     /**< Identifier.           4 +  2 = 6 */
-            uint16_t usSequenceNumber; /**< Sequence number.      6 +  2 = 8 */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xICMPEcho_IPv6 ICMPEcho_IPv6_t;
-    #endif /* ipconfigUSE_IPv6 */
-
-    #include "pack_struct_start.h"
-    struct xICMPRouterSolicitation_IPv6
-    {
-        uint8_t ucTypeOfMessage; /**<  0 +  1 =  1 */
-        uint8_t ucTypeOfService; /**<  1 +  1 =  2 */
-        uint16_t usChecksum;     /**<  2 +  2 =  4 */
-        uint32_t ulReserved;     /**<  4 +  4 =  8 */
-    }
-    #include "pack_struct_end.h"
-    typedef struct xICMPRouterSolicitation_IPv6 ICMPRouterSolicitation_IPv6_t;
-
-    #if ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 )
-        #include "pack_struct_start.h"
-        struct xICMPRouterAdvertisement_IPv6
-        {
-            uint8_t ucTypeOfMessage;       /*  0 +  1 =  1 */
-            uint8_t ucTypeOfService;       /*  1 +  1 =  2 */
-            uint16_t usChecksum;           /*  2 +  2 =  4 */
-            uint8_t ucHopLimit;            /*  4 +  1 =  5 */
-            uint8_t ucFlags;               /*  5 +  1 =  6 */
-            uint16_t usLifetime;           /*  6 +  2 =  8 */
-            uint16_t usReachableTime[ 2 ]; /*  8 +  4 = 12 */
-            uint16_t usRetransTime[ 2 ];   /* 12 +  4 = 16 */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xICMPRouterAdvertisement_IPv6 ICMPRouterAdvertisement_IPv6_t;
-    #endif /* ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 ) */
-
-    #if ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 )
-        /* This is an option with the Router Advertisement. */
-        #include "pack_struct_start.h"
-        struct xICMPPrefixOption_IPv6
-        {
-            uint8_t ucType;               /*  0 +  1 =  1 */
-            uint8_t ucLength;             /*  1 +  1 =  2 */
-            uint8_t ucPrefixLength;       /*  2 +  1 =  3 */
-            uint8_t ucFlags;              /*  3 +  1 =  4 */
-            uint32_t ulValidLifeTime;     /*  4 +  4 =  8 */
-            uint32_t ulPreferredLifeTime; /*  8 +  4 = 12 */
-            uint32_t ulReserved;          /* 12 +  4 = 16 */
-            uint8_t ucPrefix[ 16 ];       /* 16 + 16 = 32 */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xICMPPrefixOption_IPv6 ICMPPrefixOption_IPv6_t;
-    #endif /* if ( ipconfigUSE_IPv6 != 0 ) && ( ipconfigUSE_RA != 0 ) */
-
-    #include "pack_struct_start.h"
-    struct xUDP_HEADER
-    {
-        uint16_t usSourcePort;      /**< The source port                      0 + 2 = 2 */
-        uint16_t usDestinationPort; /**< The destination port                 2 + 2 = 4 */
-        uint16_t usLength;          /**< The size of the whole UDP packet     4 + 2 = 6 */
-        uint16_t usChecksum;        /**< The checksum of the whole UDP Packet 6 + 2 = 8 */
-    }
-    #include "pack_struct_end.h"
-    typedef struct xUDP_HEADER UDPHeader_t;
-
-    #include "pack_struct_start.h"
-    struct xTCP_HEADER
-    {
-        uint16_t usSourcePort;                       /**< The Source port                       0 +  2 =  2 */
-        uint16_t usDestinationPort;                  /**< The destination port                  2 +  2 =  4 */
-        uint32_t ulSequenceNumber;                   /**< The Sequence number                   4 +  4 =  8 */
-        uint32_t ulAckNr;                            /**< The acknowledgement number            8 +  4 = 12 */
-        uint8_t ucTCPOffset;                         /**< The value of TCP offset              12 +  1 = 13 */
-        uint8_t ucTCPFlags;                          /**< The TCP-flags field                  13 +  1 = 14 */
-        uint16_t usWindow;                           /**< The size of the receive window       14 +  2 = 15 */
-        uint16_t usChecksum;                         /**< The checksum of the header           15 +  2 = 18 */
-        uint16_t usUrgent;                           /**< Pointer to the last urgent data byte 18 +  2 = 20 */
-        #if ipconfigUSE_TCP == 1
-            uint8_t ucOptdata[ ipSIZE_TCP_OPTIONS ]; /**< The options + 12 = 32 */
-        #endif
-    }
-    #include "pack_struct_end.h"
-    typedef struct xTCP_HEADER TCPHeader_t;
+}
+#include "pack_struct_end.h"
+typedef struct xTCP_HEADER TCPHeader_t;
 
 /*-----------------------------------------------------------*/
 /* Nested protocol packets.                                  */
 /*-----------------------------------------------------------*/
 
-    #include "pack_struct_start.h"
-    struct xARP_PACKET
-    {
-        EthernetHeader_t xEthernetHeader; /**< The ethernet header of an ARP Packet  0 + 14 = 14 */
-        ARPHeader_t xARPHeader;           /**< The ARP header of an ARP Packet       14 + 28 = 42 */
-    }
-    #include "pack_struct_end.h"
-    typedef struct xARP_PACKET ARPPacket_t;
+#include "pack_struct_start.h"
+struct xARP_PACKET
+{
+    EthernetHeader_t xEthernetHeader;     /**< The ethernet header of an ARP Packet  0 + 14 = 14 */
+    ARPHeader_t xARPHeader;               /**< The ARP header of an ARP Packet       14 + 28 = 42 */
+}
+#include "pack_struct_end.h"
+typedef struct xARP_PACKET ARPPacket_t;
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ARPPacket_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ARPPacket_t );
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ARPPacket_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ARPPacket_t );
 
+#include "pack_struct_start.h"
+struct xIP_PACKET
+{
+    EthernetHeader_t xEthernetHeader;
+    IPHeader_t xIPHeader;
+}
+#include "pack_struct_end.h"
+typedef struct xIP_PACKET IPPacket_t;
+
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPPacket_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPPacket_t );
+
+#if ( ipconfigUSE_IPv6 != 0 )
     #include "pack_struct_start.h"
-    struct xIP_PACKET
+    struct xIP_PACKET_IPv6
     {
         EthernetHeader_t xEthernetHeader;
-        IPHeader_t xIPHeader;
+        IPHeader_IPv6_t xIPHeader;
     }
     #include "pack_struct_end.h"
-    typedef struct xIP_PACKET IPPacket_t;
+    typedef struct xIP_PACKET_IPv6 IPPacket_IPv6_t;
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPPacket_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPPacket_t );
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPPacket_IPv6_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPPacket_IPv6_t );
 
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #include "pack_struct_start.h"
-        struct xIP_PACKET_IPv6
-        {
-            EthernetHeader_t xEthernetHeader;
-            IPHeader_IPv6_t xIPHeader;
-        }
-        #include "pack_struct_end.h"
-        typedef struct xIP_PACKET_IPv6 IPPacket_IPv6_t;
+#endif /* ipconfigUSE_IPv6 */
 
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( IPPacket_IPv6_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( IPPacket_IPv6_t );
+#include "pack_struct_start.h"
+struct xICMP_PACKET
+{
+    EthernetHeader_t xEthernetHeader;     /**< The Ethernet header of an ICMP packet. */
+    IPHeader_t xIPHeader;                 /**< The IP header of an ICMP packet. */
+    ICMPHeader_t xICMPHeader;             /**< The ICMP header of an ICMP packet. */
+}
+#include "pack_struct_end.h"
+typedef struct xICMP_PACKET ICMPPacket_t;
 
-    #endif /* ipconfigUSE_IPv6 */
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ICMPPacket_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ICMPPacket_t );
 
+#if ( ipconfigUSE_IPv6 != 0 )
     #include "pack_struct_start.h"
-    struct xICMP_PACKET
+    struct xICMP_PACKET_IPv6
     {
-        EthernetHeader_t xEthernetHeader; /**< The Ethernet header of an ICMP packet. */
-        IPHeader_t xIPHeader;             /**< The IP header of an ICMP packet. */
-        ICMPHeader_t xICMPHeader;         /**< The ICMP header of an ICMP packet. */
+        EthernetHeader_t xEthernetHeader;
+        IPHeader_IPv6_t xIPHeader;
+        ICMPHeader_IPv6_t xICMPHeaderIPv6;
     }
     #include "pack_struct_end.h"
-    typedef struct xICMP_PACKET ICMPPacket_t;
+    typedef struct xICMP_PACKET_IPv6 ICMPPacket_IPv6_t;
+#endif /* if ( ipconfigUSE_IPv6 != 0 ) */
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ICMPPacket_t );
+#include "pack_struct_start.h"
+struct xUDP_PACKET
+{
+    EthernetHeader_t xEthernetHeader;     /**< UDP-Packet ethernet header  0 + 14 = 14 */
+    IPHeader_t xIPHeader;                 /**< UDP-Packet IP header        14 + 20 = 34 */
+    UDPHeader_t xUDPHeader;               /**< UDP-Packet UDP header       34 +  8 = 42 */
+}
+#include "pack_struct_end.h"
+typedef struct xUDP_PACKET UDPPacket_t;
 
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #include "pack_struct_start.h"
-        struct xICMP_PACKET_IPv6
-        {
-            EthernetHeader_t xEthernetHeader;
-            IPHeader_IPv6_t xIPHeader;
-            ICMPHeader_IPv6_t xICMPHeaderIPv6;
-        }
-        #include "pack_struct_end.h"
-        typedef struct xICMP_PACKET_IPv6 ICMPPacket_IPv6_t;
-    #endif /* if ( ipconfigUSE_IPv6 != 0 ) */
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( UDPPacket_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( UDPPacket_t );
 
+#if ( ipconfigUSE_IPv6 != 0 )
     #include "pack_struct_start.h"
-    struct xUDP_PACKET
+    struct xUDP_PACKET_IPv6
     {
-        EthernetHeader_t xEthernetHeader; /**< UDP-Packet ethernet header  0 + 14 = 14 */
-        IPHeader_t xIPHeader;             /**< UDP-Packet IP header        14 + 20 = 34 */
-        UDPHeader_t xUDPHeader;           /**< UDP-Packet UDP header       34 +  8 = 42 */
+        EthernetHeader_t xEthernetHeader;     /*  0 + 14 = 14 */
+        IPHeader_IPv6_t xIPHeader;            /* 14 + 40 = 54 */
+        UDPHeader_t xUDPHeader;               /* 54 +  8 = 62 */
     }
     #include "pack_struct_end.h"
-    typedef struct xUDP_PACKET UDPPacket_t;
+    typedef struct xUDP_PACKET_IPv6 UDPPacket_IPv6_t;
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( UDPPacket_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( UDPPacket_t );
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( UDPPacket_IPv6_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( UDPPacket_IPv6_t );
+#endif /* ipconfigUSE_IPv6 */
 
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #include "pack_struct_start.h"
-        struct xUDP_PACKET_IPv6
-        {
-            EthernetHeader_t xEthernetHeader; /*  0 + 14 = 14 */
-            IPHeader_IPv6_t xIPHeader;        /* 14 + 40 = 54 */
-            UDPHeader_t xUDPHeader;           /* 54 +  8 = 62 */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xUDP_PACKET_IPv6 UDPPacket_IPv6_t;
+#include "pack_struct_start.h"
+struct xTCP_PACKET
+{
+    EthernetHeader_t xEthernetHeader;     /**< The ethernet header  0 + 14 = 14 */
+    IPHeader_t xIPHeader;                 /**< The IP header        14 + 20 = 34 */
+    TCPHeader_t xTCPHeader;               /**< The TCP header       34 + 32 = 66 */
+}
+#include "pack_struct_end.h"
+typedef struct xTCP_PACKET TCPPacket_t;
 
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( UDPPacket_IPv6_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( UDPPacket_IPv6_t );
-    #endif /* ipconfigUSE_IPv6 */
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( TCPPacket_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( TCPPacket_t );
 
+#if ( ipconfigUSE_IPv6 != 0 )
     #include "pack_struct_start.h"
-    struct xTCP_PACKET
+    struct xTCP_PACKET_IPv6
     {
-        EthernetHeader_t xEthernetHeader; /**< The ethernet header  0 + 14 = 14 */
-        IPHeader_t xIPHeader;             /**< The IP header        14 + 20 = 34 */
-        TCPHeader_t xTCPHeader;           /**< The TCP header       34 + 32 = 66 */
+        EthernetHeader_t xEthernetHeader;     /*  0 + 14 = 14 */
+        IPHeader_IPv6_t xIPHeader;            /* 14 + 40 = 54 */
+        TCPHeader_t xTCPHeader;               /* 54 + 32 = 86 */
     }
     #include "pack_struct_end.h"
-    typedef struct xTCP_PACKET TCPPacket_t;
+    typedef struct xTCP_PACKET_IPv6 TCPPacket_IPv6_t;
+#endif /* ipconfigUSE_IPv6 */
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( TCPPacket_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( TCPPacket_t );
-
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #include "pack_struct_start.h"
-        struct xTCP_PACKET_IPv6
-        {
-            EthernetHeader_t xEthernetHeader; /*  0 + 14 = 14 */
-            IPHeader_IPv6_t xIPHeader;        /* 14 + 40 = 54 */
-            TCPHeader_t xTCPHeader;           /* 54 + 32 = 86 */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xTCP_PACKET_IPv6 TCPPacket_IPv6_t;
-    #endif /* ipconfigUSE_IPv6 */
-
-    #if ( ipconfigUSE_IPv6 != 0 )
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( TCPPacket_IPv6_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( TCPPacket_IPv6_t );
-    #endif
+#if ( ipconfigUSE_IPv6 != 0 )
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( TCPPacket_IPv6_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( TCPPacket_IPv6_t );
+#endif
 
 /**
  * Union for the protocol packet to save space. Any packet cannot have more than one
  * of the below protocol packets.
  */
-    typedef union xPROT_PACKET
-    {
-        ARPPacket_t xARPPacket;   /**< Union member: ARP packet struct */
-        TCPPacket_t xTCPPacket;   /**< Union member: TCP packet struct */
-        UDPPacket_t xUDPPacket;   /**< Union member: UDP packet struct */
-        ICMPPacket_t xICMPPacket; /**< Union member: ICMP packet struct */
-    } ProtocolPacket_t;
+typedef union xPROT_PACKET
+{
+    ARPPacket_t xARPPacket;       /**< Union member: ARP packet struct */
+    TCPPacket_t xTCPPacket;       /**< Union member: TCP packet struct */
+    UDPPacket_t xUDPPacket;       /**< Union member: UDP packet struct */
+    ICMPPacket_t xICMPPacket;     /**< Union member: ICMP packet struct */
+} ProtocolPacket_t;
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ProtocolPacket_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ProtocolPacket_t );
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ProtocolPacket_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ProtocolPacket_t );
 
 /**
  * Union for protocol headers to save space (RAM). Any packet cannot have more than one of
  * the below protocols.
  */
-    typedef union xPROT_HEADERS
-    {
-        ICMPHeader_t xICMPHeader;              /**< Union member: ICMP header */
-        UDPHeader_t xUDPHeader;                /**< Union member: UDP header */
-        TCPHeader_t xTCPHeader;                /**< Union member: TCP header */
-        #if ( ipconfigUSE_IPv6 != 0 )
-            ICMPHeader_IPv6_t xICMPHeaderIPv6; /**< Union member: ICMPv6 header */
-        #endif
-    } ProtocolHeaders_t;
+typedef union xPROT_HEADERS
+{
+    ICMPHeader_t xICMPHeader;                  /**< Union member: ICMP header */
+    UDPHeader_t xUDPHeader;                    /**< Union member: UDP header */
+    TCPHeader_t xTCPHeader;                    /**< Union member: TCP header */
+    #if ( ipconfigUSE_IPv6 != 0 )
+        ICMPHeader_IPv6_t xICMPHeaderIPv6;     /**< Union member: ICMPv6 header */
+    #endif
+} ProtocolHeaders_t;
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ProtocolHeaders_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ProtocolHeaders_t );
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ProtocolHeaders_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ProtocolHeaders_t );
 
 /** @brief This struct describes a packet, it is used by the function
  * usGenerateProtocolChecksum(). */
-    struct xPacketSummary
-    {
-        #if ( ipconfigUSE_IPv6 != 0 )
-            BaseType_t xIsIPv6;                      /**< pdTRUE for IPv6 packets. */
-            const IPHeader_IPv6_t * pxIPPacket_IPv6; /**< A pointer to the IPv6 header. */
-        #endif
-        #if ( ipconfigHAS_DEBUG_PRINTF != 0 )
-            const char * pcType;               /**< Just for logging purposes: the name of the protocol. */
-        #endif
-        size_t uxIPHeaderLength;               /**< Either 40 or 20, depending on the IP-type */
-        size_t uxProtocolHeaderLength;         /**< Either 8, 20, or more or 20, depending on the protocol-type */
-        uint16_t usChecksum;                   /**< Checksum accumulator. */
-        uint8_t ucProtocol;                    /**< ipPROTOCOL_TCP, ipPROTOCOL_UDP, ipPROTOCOL_ICMP */
-        const IPPacket_t * pxIPPacket;         /**< A pointer to the IPv4 header. */
-        ProtocolHeaders_t * pxProtocolHeaders; /**< Points to first byte after IP-header */
-        uint16_t usPayloadLength;              /**< Property of IP-header (for IPv4: length of IP-header included) */
-        uint16_t usProtocolBytes;              /**< The total length of the protocol data. */
-        uint16_t * pusChecksum;                /**< A pointer to the location where the protocol checksum is stored. */
-    };
+struct xPacketSummary
+{
+    #if ( ipconfigUSE_IPv6 != 0 )
+        BaseType_t xIsIPv6;                          /**< pdTRUE for IPv6 packets. */
+        const IPHeader_IPv6_t * pxIPPacket_IPv6;     /**< A pointer to the IPv6 header. */
+    #endif
+    #if ( ipconfigHAS_DEBUG_PRINTF != 0 )
+        const char * pcType;                   /**< Just for logging purposes: the name of the protocol. */
+    #endif
+    size_t uxIPHeaderLength;                   /**< Either 40 or 20, depending on the IP-type */
+    size_t uxProtocolHeaderLength;             /**< Either 8, 20, or more or 20, depending on the protocol-type */
+    uint16_t usChecksum;                       /**< Checksum accumulator. */
+    uint8_t ucProtocol;                        /**< ipPROTOCOL_TCP, ipPROTOCOL_UDP, ipPROTOCOL_ICMP */
+    const IPPacket_t * pxIPPacket;             /**< A pointer to the IPv4 header. */
+    ProtocolHeaders_t * pxProtocolHeaders;     /**< Points to first byte after IP-header */
+    uint16_t usPayloadLength;                  /**< Property of IP-header (for IPv4: length of IP-header included) */
+    uint16_t usProtocolBytes;                  /**< The total length of the protocol data. */
+    uint16_t * pusChecksum;                    /**< A pointer to the location where the protocol checksum is stored. */
+};
 
 /* The maximum UDP payload length. */
-    #if ( ipconfigUSE_IPv6 != 0 )
-        #define ipMAX_UDP_PAYLOAD_LENGTH    ( ( ipconfigNETWORK_MTU - ipSIZE_OF_IPv6_HEADER ) - ipSIZE_OF_UDP_HEADER )
-    #else
-        #define ipMAX_UDP_PAYLOAD_LENGTH    ( ( ipconfigNETWORK_MTU - ipSIZE_OF_IPv4_HEADER ) - ipSIZE_OF_UDP_HEADER )
-    #endif
+#if ( ipconfigUSE_IPv6 != 0 )
+    #define ipMAX_UDP_PAYLOAD_LENGTH    ( ( ipconfigNETWORK_MTU - ipSIZE_OF_IPv6_HEADER ) - ipSIZE_OF_UDP_HEADER )
+#else
+    #define ipMAX_UDP_PAYLOAD_LENGTH    ( ( ipconfigNETWORK_MTU - ipSIZE_OF_IPv4_HEADER ) - ipSIZE_OF_UDP_HEADER )
+#endif
 
-    typedef enum
-    {
-        eReleaseBuffer = 0,   /* Processing the frame did not find anything to do - just release the buffer. */
-        eProcessBuffer,       /* An Ethernet frame has a valid address - continue process its contents. */
-        eReturnEthernetFrame, /* The Ethernet frame contains an ARP or ICMP packet that can be returned to its source. */
-        eFrameConsumed,       /* Processing the Ethernet packet contents resulted in the payload being sent to the stack. */
-        eWaitingARPResolution /* Frame is awaiting ARP resolution. */
-    } eFrameProcessingResult_t;
+typedef enum
+{
+    eReleaseBuffer = 0,       /* Processing the frame did not find anything to do - just release the buffer. */
+    eProcessBuffer,           /* An Ethernet frame has a valid address - continue process its contents. */
+    eReturnEthernetFrame,     /* The Ethernet frame contains an ARP or ICMP packet that can be returned to its source. */
+    eFrameConsumed,           /* Processing the Ethernet packet contents resulted in the payload being sent to the stack. */
+    eWaitingARPResolution     /* Frame is awaiting ARP resolution. */
+} eFrameProcessingResult_t;
 
-    typedef enum
-    {
-        eNoEvent = -1,
-        eNetworkDownEvent,  /* 0: The network interface has been lost and/or needs [re]connecting. */
-        eNetworkRxEvent,    /* 1: The network interface has queued a received Ethernet frame. */
-        eNetworkTxEvent,    /* 2: Let the IP-task send a network packet. */
-        eARPTimerEvent,     /* 3: The ARP timer expired. */
-        eStackTxEvent,      /* 4: The software stack has queued a packet to transmit. */
-        eDHCP_RA_Event,     /* 5: Process the DHCP or RA/SLAAC state machine. */
-        eTCPTimerEvent,     /* 6: See if any TCP socket needs attention. */
-        eTCPAcceptEvent,    /* 7: Client API FreeRTOS_accept() waiting for client connections. */
-        eTCPNetStat,        /* 8: IP-task is asked to produce a netstat listing. */
-        eSocketBindEvent,   /* 9: Send a message to the IP-task to bind a socket to a port. */
-        eSocketCloseEvent,  /*10: Send a message to the IP-task to close a socket. */
-        eSocketSelectEvent, /*11: Send a message to the IP-task for select(). */
-        eSocketSignalEvent  /*12: A socket must be signalled. */
-    } eIPEvent_t;
+typedef enum
+{
+    eNoEvent = -1,
+    eNetworkDownEvent,      /* 0: The network interface has been lost and/or needs [re]connecting. */
+    eNetworkRxEvent,        /* 1: The network interface has queued a received Ethernet frame. */
+    eNetworkTxEvent,        /* 2: Let the IP-task send a network packet. */
+    eARPTimerEvent,         /* 3: The ARP timer expired. */
+    eStackTxEvent,          /* 4: The software stack has queued a packet to transmit. */
+    eDHCP_RA_Event,         /* 5: Process the DHCP or RA/SLAAC state machine. */
+    eTCPTimerEvent,         /* 6: See if any TCP socket needs attention. */
+    eTCPAcceptEvent,        /* 7: Client API FreeRTOS_accept() waiting for client connections. */
+    eTCPNetStat,            /* 8: IP-task is asked to produce a netstat listing. */
+    eSocketBindEvent,       /* 9: Send a message to the IP-task to bind a socket to a port. */
+    eSocketCloseEvent,      /*10: Send a message to the IP-task to close a socket. */
+    eSocketSelectEvent,     /*11: Send a message to the IP-task for select(). */
+    eSocketSignalEvent      /*12: A socket must be signalled. */
+} eIPEvent_t;
 
 /**
  * Structure for the information of the commands issued to the IP task.
  */
-    typedef struct IP_TASK_COMMANDS
-    {
-        eIPEvent_t eEventType; /**< The event-type enum */
-        void * pvData;         /**< The data in the event */
-    } IPStackEvent_t;
+typedef struct IP_TASK_COMMANDS
+{
+    eIPEvent_t eEventType;     /**< The event-type enum */
+    void * pvData;             /**< The data in the event */
+} IPStackEvent_t;
 
-    #define ipBROADCAST_IP_ADDRESS    0xffffffffU
+#define ipBROADCAST_IP_ADDRESS    0xffffffffU
 
 /** @brief The pointer to buffer with packet waiting for ARP resolution. This variable
  *  is defined in FreeRTOS_IP.c.
  * Note that the variable is private to the library. */
-    extern NetworkBufferDescriptor_t * pxARPWaitingNetworkBuffer;
+extern NetworkBufferDescriptor_t * pxARPWaitingNetworkBuffer;
 
 /* Offset into the Ethernet frame that is used to temporarily store information
  * on the fragmentation status of the packet being sent.  The value is important,
  * as it is past the location into which the destination address will get placed. */
-    #define ipFRAGMENTATION_PARAMETERS_OFFSET    ( 6 )
-    #define ipSOCKET_OPTIONS_OFFSET              ( 6 )
+#define ipFRAGMENTATION_PARAMETERS_OFFSET    ( 6 )
+#define ipSOCKET_OPTIONS_OFFSET              ( 6 )
 
 /* The offset into a UDP packet at which the UDP data (payload) starts. */
-    #define ipUDP_PAYLOAD_OFFSET_IPv4            ( sizeof( UDPPacket_t ) )
+#define ipUDP_PAYLOAD_OFFSET_IPv4            ( sizeof( UDPPacket_t ) )
 
 /* The offset into a UDP packet at which the UDP data (payload) starts. */
-    #define ipUDP_PAYLOAD_OFFSET_IPv6            ( sizeof( UDPPacket_IPv6_t ) )
+#define ipUDP_PAYLOAD_OFFSET_IPv6            ( sizeof( UDPPacket_IPv6_t ) )
 
 /* The offset from the UDP payload where the IP type will be stored.
  * For IPv4 packets, this it located 6 bytes before pucEthernetBuffer.
  * For IPv6 packets, this it located in the usual 'ucVersionTrafficClass'. */
-    #define ipIP_TYPE_OFFSET                     ( 6U )
+#define ipIP_TYPE_OFFSET                     ( 6U )
 /* The value of 'ipUDP_PAYLOAD_IP_TYPE_OFFSET' is 42 + 6 = 48 bytes. */
-    #define ipUDP_PAYLOAD_IP_TYPE_OFFSET         ( sizeof( UDPPacket_t ) + ipIP_TYPE_OFFSET )
+#define ipUDP_PAYLOAD_IP_TYPE_OFFSET         ( sizeof( UDPPacket_t ) + ipIP_TYPE_OFFSET )
 
 /* The offset into an IP packet into which the IP data (payload) starts. */
-    #define ipIP_PAYLOAD_OFFSET                  ( sizeof( IPPacket_t ) )
+#define ipIP_PAYLOAD_OFFSET                  ( sizeof( IPPacket_t ) )
 
-    #if ( ipconfigBYTE_ORDER == pdFREERTOS_LITTLE_ENDIAN )
-
-/* Ethernet frame types. */
-        #define ipARP_FRAME_TYPE                ( 0x0608U )
-        #define ipIPv4_FRAME_TYPE               ( 0x0008U )
-        #define ipIPv6_FRAME_TYPE               ( 0xDD86U )
-
-/* ARP related definitions. */
-        #define ipARP_PROTOCOL_TYPE             ( 0x0008U )
-        #define ipARP_HARDWARE_TYPE_ETHERNET    ( 0x0100U )
-        #define ipARP_REQUEST                   ( 0x0100U )
-        #define ipARP_REPLY                     ( 0x0200U )
-
-    #else /* if ( ipconfigBYTE_ORDER == pdFREERTOS_LITTLE_ENDIAN ) */
+#if ( ipconfigBYTE_ORDER == pdFREERTOS_LITTLE_ENDIAN )
 
 /* Ethernet frame types. */
-        #define ipARP_FRAME_TYPE                ( 0x0806U )
-        #define ipIPv4_FRAME_TYPE               ( 0x0800U )
-        #define ipIPv6_FRAME_TYPE               ( 0x86DDU )
+    #define ipARP_FRAME_TYPE                ( 0x0608U )
+    #define ipIPv4_FRAME_TYPE               ( 0x0008U )
+    #define ipIPv6_FRAME_TYPE               ( 0xDD86U )
 
 /* ARP related definitions. */
-        #define ipARP_PROTOCOL_TYPE             ( 0x0800U )
-        #define ipARP_HARDWARE_TYPE_ETHERNET    ( 0x0001U )
-        #define ipARP_REQUEST                   ( 0x0001 )
-        #define ipARP_REPLY                     ( 0x0002 )
+    #define ipARP_PROTOCOL_TYPE             ( 0x0008U )
+    #define ipARP_HARDWARE_TYPE_ETHERNET    ( 0x0100U )
+    #define ipARP_REQUEST                   ( 0x0100U )
+    #define ipARP_REPLY                     ( 0x0200U )
 
-    #endif /* ipconfigBYTE_ORDER == pdFREERTOS_LITTLE_ENDIAN */
+#else /* if ( ipconfigBYTE_ORDER == pdFREERTOS_LITTLE_ENDIAN ) */
 
-    #if ( ipconfigDNS_USE_CALLBACKS != 0 )
-        /* Two functions that are only called from within the library. */
-        extern void vDNSInitialise( void );
-        extern void vDNSCheckCallBack( void * pvSearchID );
-    #endif /* ipconfigDNS_USE_CALLBACKS */
+/* Ethernet frame types. */
+    #define ipARP_FRAME_TYPE                ( 0x0806U )
+    #define ipIPv4_FRAME_TYPE               ( 0x0800U )
+    #define ipIPv6_FRAME_TYPE               ( 0x86DDU )
+
+/* ARP related definitions. */
+    #define ipARP_PROTOCOL_TYPE             ( 0x0800U )
+    #define ipARP_HARDWARE_TYPE_ETHERNET    ( 0x0001U )
+    #define ipARP_REQUEST                   ( 0x0001 )
+    #define ipARP_REPLY                     ( 0x0002 )
+
+#endif /* ipconfigBYTE_ORDER == pdFREERTOS_LITTLE_ENDIAN */
 
 /* For convenience, a MAC address of all zeros and another of all 0xffs are
  * defined const for quick reference. */
-    extern const MACAddress_t xBroadcastMACAddress; /* all 0xff's */
-    extern uint16_t usPacketIdentifier;
+extern const MACAddress_t xBroadcastMACAddress;     /* all 0xff's */
+extern uint16_t usPacketIdentifier;
 
 /** @brief The list that contains mappings between sockets and port numbers.
  *         Accesses to this list must be protected by critical sections of
  *         some kind.
  */
-    extern List_t xBoundUDPSocketsList;
+extern List_t xBoundUDPSocketsList;
 
-    #if ipconfigUSE_TCP == 1
+#if ipconfigUSE_TCP == 1
 
 /** @brief The list that contains mappings between sockets and port numbers.
  *         Accesses to this list must be protected by critical sections of
  *         some kind.
  */
-        extern List_t xBoundTCPSocketsList;
+    extern List_t xBoundTCPSocketsList;
 
-    #endif /* ipconfigUSE_TCP == 1 */
+#endif /* ipconfigUSE_TCP == 1 */
 
 /* True when BufferAllocation_1.c was included, false for BufferAllocation_2.c */
-    extern const BaseType_t xBufferAllocFixedSize;
+extern const BaseType_t xBufferAllocFixedSize;
 
 /* As FreeRTOS_Routing is included later, use forward declarations
  * of the two structs. */
-    struct xNetworkEndPoint;
-    struct xNetworkInterface;
+struct xNetworkEndPoint;
+struct xNetworkInterface;
 
 /* A list of all network end-points: */
-    extern struct xNetworkEndPoint * pxNetworkEndPoints;
+extern struct xNetworkEndPoint * pxNetworkEndPoints;
 
 /* A list of all network interfaces: */
-    extern struct xNetworkInterface * pxNetworkInterfaces;
+extern struct xNetworkInterface * pxNetworkInterfaces;
 
 /* In this library, there is often a cast from a character pointer
  * to a pointer to a struct.
@@ -598,7 +601,7 @@
  * The MISRA warnings can safely be suppressed because all casts are planned with care.
  */
 
-    #define ipPOINTER_CAST( TYPE, pointer )    ( ( TYPE ) ( pointer ) )
+#define ipPOINTER_CAST( TYPE, pointer )    ( ( TYPE ) ( pointer ) )
 
 /* Sequence and ACK numbers are essentially unsigned (uint32_t). But when
  * a distance is calculated, it is useful to use signed numbers:
@@ -611,66 +614,66 @@
  * -emacro(9030,ipNUMERIC_CAST) // 9030: Impermissible cast; cannot cast from 'essentially Boolean' to 'essentially signed' [MISRA 2012 Rule 10.5, advisory])
  */
 
-    #define ipNUMERIC_CAST( TYPE, expression )    ( ( TYPE ) ( expression ) )
+#define ipNUMERIC_CAST( TYPE, expression )    ( ( TYPE ) ( expression ) )
 
 /* ICMP packets are sent using the same function as UDP packets.  The port
  * number is used to distinguish between the two, as 0 is an invalid UDP port. */
-    #define ipPACKET_CONTAINS_ICMP_DATA    ( 0 )
+#define ipPACKET_CONTAINS_ICMP_DATA    ( 0 )
 
 /* For now, the lower 8 bits in 'xEventBits' will be reserved for the above
  * socket events. */
-    #define SOCKET_EVENT_BIT_COUNT         8
+#define SOCKET_EVENT_BIT_COUNT         8
 
 /** @brief The macros vSetField16() and vSetField32() will write either a short or a 32-bit
  * value into an array of bytes. They will be stored big-endian.
  * The helper functions do the actual work.
  */
-    extern void vSetField16helper( uint8_t * pucBase,
-                                   size_t uxOffset,
-                                   uint16_t usValue );
-    #define vSetField16( pucBase, xType, xField, usValue ) \
+extern void vSetField16helper( uint8_t * pucBase,
+                               size_t uxOffset,
+                               uint16_t usValue );
+#define vSetField16( pucBase, xType, xField, usValue ) \
     vSetField16helper( pucBase, offsetof( xType, xField ), usValue )
 
-    extern void vSetField32helper( uint8_t * pucBase,
-                                   size_t uxOffset,
-                                   uint32_t ulValue );
-    #define vSetField32( pucBase, xType, xField, ulValue ) \
+extern void vSetField32helper( uint8_t * pucBase,
+                               size_t uxOffset,
+                               uint32_t ulValue );
+#define vSetField32( pucBase, xType, xField, ulValue ) \
     vSetField32helper( pucBase, offsetof( xType, xField ), ulValue )
 
-    #define vFlip_16( left, right ) \
-    do {                            \
-        uint16_t tmp = ( left );    \
-        ( left ) = ( right );       \
-        ( right ) = tmp;            \
+#define vFlip_16( left, right )  \
+    do {                         \
+        uint16_t tmp = ( left ); \
+        ( left ) = ( right );    \
+        ( right ) = tmp;         \
     } while( ipFALSE_BOOL )
 
-    #define vFlip_32( left, right ) \
-    do {                            \
-        uint32_t tmp = ( left );    \
-        ( left ) = ( right );       \
-        ( right ) = tmp;            \
+#define vFlip_32( left, right )  \
+    do {                         \
+        uint32_t tmp = ( left ); \
+        ( left ) = ( right );    \
+        ( right ) = tmp;         \
     } while( ipFALSE_BOOL )
 
 /* WARNING: Do NOT use this macro when the array was received as a parameter. */
-    #ifndef ARRAY_SIZE
-        /** @brief Macro calculates the number of elements in an array as a BaseType_t. */
-        #define ARRAY_SIZE( x )    ( ( BaseType_t ) ( sizeof( x ) / sizeof( ( x )[ 0 ] ) ) )
-    #endif
+#ifndef ARRAY_SIZE
+    /** @brief Macro calculates the number of elements in an array as a BaseType_t. */
+    #define ARRAY_SIZE( x )    ( ( BaseType_t ) ( sizeof( x ) / sizeof( ( x )[ 0 ] ) ) )
+#endif
 
-    #ifndef _WINDOWS_
-        /** @brief Macro calculates the number of elements in an array as a size_t. */
-        #ifndef ARRAY_SIZE_X
-            #define ARRAY_SIZE_X( x )                        \
+#ifndef _WINDOWS_
+    /** @brief Macro calculates the number of elements in an array as a size_t. */
+    #ifndef ARRAY_SIZE_X
+        #define ARRAY_SIZE_X( x )                            \
     ( { size_t uxCount = ( sizeof( x ) / sizeof( x )[ 0 ] ); \
         BaseType_t xCount = ( BaseType_t ) uxCount;          \
         xCount; }                                            \
     )
-        #endif
-    #else
-        #ifndef ARRAY_SIZE_X
-            #define ARRAY_SIZE_X    ARRAY_SIZE
-        #endif
-    #endif /* ifndef _WINDOWS_ */
+    #endif
+#else
+    #ifndef ARRAY_SIZE_X
+        #define ARRAY_SIZE_X    ARRAY_SIZE
+    #endif
+#endif /* ifndef _WINDOWS_ */
 
 /*
  * Create a message that contains a command to initialise the network interface.
@@ -684,27 +687,27 @@
  * returns a non-zero value then a context switch should be performed before
  * the interrupt is exited.
  */
-    void FreeRTOS_NetworkDown( struct xNetworkInterface * pxNetworkInterface );
-    BaseType_t FreeRTOS_NetworkDownFromISR( struct xNetworkInterface * pxNetworkInterface );
+void FreeRTOS_NetworkDown( struct xNetworkInterface * pxNetworkInterface );
+BaseType_t FreeRTOS_NetworkDownFromISR( struct xNetworkInterface * pxNetworkInterface );
 
 /*
  * Processes incoming ARP packets.
  */
-    eFrameProcessingResult_t eARPProcessPacket( NetworkBufferDescriptor_t * pxNetworkBuffer );
+eFrameProcessingResult_t eARPProcessPacket( NetworkBufferDescriptor_t * pxNetworkBuffer );
 
 /*
  * Inspect an Ethernet frame to see if it contains data that the stack needs to
  * process.  eProcessBuffer is returned if the frame should be processed by the
  * stack.  eReleaseBuffer is returned if the frame should be discarded.
  */
-    eFrameProcessingResult_t eConsiderFrameForProcessing( const uint8_t * const pucEthernetBuffer );
+eFrameProcessingResult_t eConsiderFrameForProcessing( const uint8_t * const pucEthernetBuffer );
 
 /*
  * Return the checksum generated over xDataLengthBytes from pucNextData.
  */
-    uint16_t usGenerateChecksum( uint16_t usSum,
-                                 const uint8_t * pucNextData,
-                                 size_t uxByteCount );
+uint16_t usGenerateChecksum( uint16_t usSum,
+                             const uint8_t * pucNextData,
+                             size_t uxByteCount );
 
 /* Socket related private functions. */
 
@@ -713,40 +716,40 @@
  * payload size (excluding packet headers) and that the packet in pucEthernetBuffer
  * is at least the size of UDPPacket_t.
  */
-    BaseType_t xProcessReceivedUDPPacket( NetworkBufferDescriptor_t * pxNetworkBuffer,
-                                          uint16_t usPort,
-                                          BaseType_t * pxIsWaitingARPResolution );
+BaseType_t xProcessReceivedUDPPacket( NetworkBufferDescriptor_t * pxNetworkBuffer,
+                                      uint16_t usPort,
+                                      BaseType_t * pxIsWaitingARPResolution );
 
 /*
  * Initialize the socket list data structures for TCP and UDP.
  */
-    void vNetworkSocketsInit( void );
+void vNetworkSocketsInit( void );
 
 /*
  * Returns pdTRUE if the IP task has been created and is initialised.  Otherwise
  * returns pdFALSE.
  */
-    BaseType_t xIPIsNetworkTaskReady( void );
+BaseType_t xIPIsNetworkTaskReady( void );
 
-    #if ( ipconfigSOCKET_HAS_USER_WAKE_CALLBACK == 1 )
-        struct xSOCKET;
-        typedef void (* SocketWakeupCallback_t)( struct xSOCKET * pxSocket );
-    #endif
+#if ( ipconfigSOCKET_HAS_USER_WAKE_CALLBACK == 1 )
+    struct xSOCKET;
+    typedef void (* SocketWakeupCallback_t)( struct xSOCKET * pxSocket );
+#endif
 
-    #if ( ipconfigUSE_TCP == 1 )
+#if ( ipconfigUSE_TCP == 1 )
 
 /*
  * Actually a user thing, but because xBoundTCPSocketsList, let it do by the
  * IP-task
  */
-        #if ( ipconfigHAS_PRINTF != 0 )
-            void vTCPNetStat( void );
-        #endif
+    #if ( ipconfigHAS_PRINTF != 0 )
+        void vTCPNetStat( void );
+    #endif
 
 /*
  * At least one socket needs to check for timeouts
  */
-        TickType_t xTCPTimerCheck( BaseType_t xWillSleep );
+    TickType_t xTCPTimerCheck( BaseType_t xWillSleep );
 
 /**
  * Every TCP socket has a buffer space just big enough to store
@@ -754,24 +757,24 @@
  * As a reference of this field may be passed to DMA, force the
  * alignment to 8 bytes.
  */
-        typedef union
+    typedef union
+    {
+        struct
         {
-            struct
-            {
-                uint64_t ullAlignmentWord; /**< Increase the alignment of this union by adding a 64-bit variable. */
-            } a;                           /**< A struct to increase alignment. */
-            struct
-            {
-                /* The next field only serves to give 'ucLastPacket' a correct
-                 * alignment of 8 + 2.  See comments in FreeRTOS_IP.h */
-                uint8_t ucFillPacket[ ipconfigPACKET_FILLER_SIZE ];
-                #if ( ipconfigUSE_IPv6 != 0 )
-                    uint8_t ucLastPacket[ sizeof( TCPPacket_IPv6_t ) ];
-                #else
-                    uint8_t ucLastPacket[ sizeof( TCPPacket_t ) ];
-                #endif
-            } u; /**< The structure to give an alignment of 8 + 2 */
-        } LastTCPPacket_t;
+            uint64_t ullAlignmentWord;     /**< Increase the alignment of this union by adding a 64-bit variable. */
+        } a;                               /**< A struct to increase alignment. */
+        struct
+        {
+            /* The next field only serves to give 'ucLastPacket' a correct
+             * alignment of 8 + 2.  See comments in FreeRTOS_IP.h */
+            uint8_t ucFillPacket[ ipconfigPACKET_FILLER_SIZE ];
+            #if ( ipconfigUSE_IPv6 != 0 )
+                uint8_t ucLastPacket[ sizeof( TCPPacket_IPv6_t ) ];
+            #else
+                uint8_t ucLastPacket[ sizeof( TCPPacket_t ) ];
+            #endif
+        } u;     /**< The structure to give an alignment of 8 + 2 */
+    } LastTCPPacket_t;
 
 /**
  * Note that the values of all short and long integers in these structs
@@ -779,232 +782,232 @@
  * Translation should take place when accessing any structure which defines
  * network packets, such as IPHeader_t and TCPHeader_t
  */
-        typedef struct TCPSOCKET
+    typedef struct TCPSOCKET
+    {
+        uint32_t ulRemoteIP;     /**< IP address of remote machine */
+        #if ( ipconfigUSE_IPv6 != 0 )
+            IPv6_Address_t xRemoteIP_IPv6;
+        #endif
+        uint16_t usRemotePort;     /**< Port on remote machine */
+        struct
         {
-            uint32_t ulRemoteIP; /**< IP address of remote machine */
-            #if ( ipconfigUSE_IPv6 != 0 )
-                IPv6_Address_t xRemoteIP_IPv6;
-            #endif
-            uint16_t usRemotePort; /**< Port on remote machine */
-            struct
-            {
-                /* Most compilers do like bit-flags */
-                uint32_t
-                    bMssChange : 1,        /**< This socket has seen a change in MSS */
-                    bPassAccept : 1,       /**< when true, this socket may be returned in a call to accept() */
-                    bPassQueued : 1,       /**< when true, this socket is an orphan until it gets connected
+            /* Most compilers do like bit-flags */
+            uint32_t
+                bMssChange : 1,            /**< This socket has seen a change in MSS */
+                bPassAccept : 1,           /**< when true, this socket may be returned in a call to accept() */
+                bPassQueued : 1,           /**< when true, this socket is an orphan until it gets connected
                                             * Why an orphan? Because it may not be returned in a accept() call until it
                                             * gets the state eESTABLISHED */
-                    bReuseSocket : 1,      /**< When a listening socket gets a connection, do not create a new instance but keep on using it */
-                    bCloseAfterSend : 1,   /**< As soon as the last byte has been transmitted, finalise the connection
+                bReuseSocket : 1,          /**< When a listening socket gets a connection, do not create a new instance but keep on using it */
+                bCloseAfterSend : 1,       /**< As soon as the last byte has been transmitted, finalise the connection
                                             * Useful in e.g. FTP connections, where the last data bytes are sent along with the FIN flag */
-                    bUserShutdown : 1,     /**< User requesting a graceful shutdown */
-                    bCloseRequested : 1,   /**< Request to finalise the connection */
-                    bLowWater : 1,         /**< high-water level has been reached. Cleared as soon as 'rx-count < lo-water' */
-                    bWinChange : 1,        /**< The value of bLowWater has changed, must send a window update */
-                    bSendKeepAlive : 1,    /**< When this flag is true, a TCP keep-alive message must be send */
-                    bWaitKeepAlive : 1,    /**< When this flag is true, a TCP keep-alive reply is expected */
-                    bConnPrepared : 1,     /**< Connecting socket: Message has been prepared */
-                #if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 )
-                    bConnPassed : 1,       /**< Connecting socket: Socket has been passed in a successful select()  */
-                #endif /* ipconfigSUPPORT_SELECT_FUNCTION */
-                bFinAccepted : 1,          /**< This socket has received (or sent) a FIN and accepted it */
-                    bFinSent : 1,          /**< We've sent out a FIN */
-                    bFinRecv : 1,          /**< We've received a FIN from our peer */
-                    bFinAcked : 1,         /**< Our FIN packet has been acked */
-                    bFinLast : 1,          /**< The last ACK (after FIN and FIN+ACK) has been sent or will be sent by the peer */
-                    bRxStopped : 1,        /**< Application asked to temporarily stop reception */
-                    bMallocError : 1,      /**< There was an error allocating a stream */
-                    bWinScaling : 1;       /**< A TCP-Window Scaling option was offered and accepted in the SYN phase. */
-            } bits;                        /**< The bits structure */
-            uint32_t ulHighestRxAllowed;   /**< The highest sequence number that we can receive at any moment */
-            uint16_t usTimeout;            /**< Time (in ticks) after which this socket needs attention */
-            uint16_t usMSS;                /**< The Maximum Segment Size for the current connection. */
-            uint16_t usChildCount;         /**< In case of a listening socket: number of connections on this port number */
-            uint16_t usBacklog;            /**< In case of a listening socket: maximum number of concurrent connections on this port number */
-            uint8_t ucRepCount;            /**< Send repeat count, for retransmissions
+                bUserShutdown : 1,         /**< User requesting a graceful shutdown */
+                bCloseRequested : 1,       /**< Request to finalise the connection */
+                bLowWater : 1,             /**< high-water level has been reached. Cleared as soon as 'rx-count < lo-water' */
+                bWinChange : 1,            /**< The value of bLowWater has changed, must send a window update */
+                bSendKeepAlive : 1,        /**< When this flag is true, a TCP keep-alive message must be send */
+                bWaitKeepAlive : 1,        /**< When this flag is true, a TCP keep-alive reply is expected */
+                bConnPrepared : 1,         /**< Connecting socket: Message has been prepared */
+            #if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 )
+                bConnPassed : 1,           /**< Connecting socket: Socket has been passed in a successful select()  */
+            #endif /* ipconfigSUPPORT_SELECT_FUNCTION */
+            bFinAccepted : 1,              /**< This socket has received (or sent) a FIN and accepted it */
+                bFinSent : 1,              /**< We've sent out a FIN */
+                bFinRecv : 1,              /**< We've received a FIN from our peer */
+                bFinAcked : 1,             /**< Our FIN packet has been acked */
+                bFinLast : 1,              /**< The last ACK (after FIN and FIN+ACK) has been sent or will be sent by the peer */
+                bRxStopped : 1,            /**< Application asked to temporarily stop reception */
+                bMallocError : 1,          /**< There was an error allocating a stream */
+                bWinScaling : 1;           /**< A TCP-Window Scaling option was offered and accepted in the SYN phase. */
+        } bits;                            /**< The bits structure */
+        uint32_t ulHighestRxAllowed;       /**< The highest sequence number that we can receive at any moment */
+        uint16_t usTimeout;                /**< Time (in ticks) after which this socket needs attention */
+        uint16_t usMSS;                    /**< The Maximum Segment Size for the current connection. */
+        uint16_t usChildCount;             /**< In case of a listening socket: number of connections on this port number */
+        uint16_t usBacklog;                /**< In case of a listening socket: maximum number of concurrent connections on this port number */
+        uint8_t ucRepCount;                /**< Send repeat count, for retransmissions
                                             * This counter is separate from the xmitCount in the
                                             * TCP win segments */
-            uint8_t ucTCPState;            /**< TCP state: see eTCP_STATE */
-            struct xSOCKET * pxPeerSocket; /**< for server socket: child, for child socket: parent */
-            #if ( ipconfigTCP_KEEP_ALIVE == 1 )
-                uint8_t ucKeepRepCount;
-                TickType_t xLastAliveTime; /**< The last value of keepalive time.*/
-            #endif /* ipconfigTCP_KEEP_ALIVE */
-            #if ( ipconfigTCP_HANG_PROTECTION == 1 )
-                TickType_t xLastActTime;                  /**< The last time when hang-protection was done.*/
-            #endif /* ipconfigTCP_HANG_PROTECTION */
-            size_t uxLittleSpace;                         /**< The value deemed as low amount of space. */
-            size_t uxEnoughSpace;                         /**< The value deemed as enough space. */
-            size_t uxRxStreamSize;                        /**< The Receive stream size */
-            size_t uxTxStreamSize;                        /**< The transmit stream size */
-            StreamBuffer_t * rxStream;                    /**< The pointer to the receive stream buffer. */
-            StreamBuffer_t * txStream;                    /**< The pointer to the transmit stream buffer. */
-            #if ( ipconfigUSE_TCP_WIN == 1 )
-                NetworkBufferDescriptor_t * pxAckMessage; /**< The pointer to the ACK message */
-            #endif /* ipconfigUSE_TCP_WIN */
-            LastTCPPacket_t xPacket;                      /**< Buffer space to store the last TCP header received. */
-            uint8_t tcpflags;                             /**< TCP flags */
-            #if ( ipconfigUSE_TCP_WIN != 0 )
-                uint8_t ucMyWinScaleFactor;               /**< Scaling factor of this device. */
-                uint8_t ucPeerWinScaleFactor;             /**< Scaling factor of the peer. */
-            #endif
-            #if ( ipconfigUSE_CALLBACKS == 1 )
-                FOnTCPReceive_t pxHandleReceive;  /**<
+        uint8_t ucTCPState;                /**< TCP state: see eTCP_STATE */
+        struct xSOCKET * pxPeerSocket;     /**< for server socket: child, for child socket: parent */
+        #if ( ipconfigTCP_KEEP_ALIVE == 1 )
+            uint8_t ucKeepRepCount;
+            TickType_t xLastAliveTime;     /**< The last value of keepalive time.*/
+        #endif /* ipconfigTCP_KEEP_ALIVE */
+        #if ( ipconfigTCP_HANG_PROTECTION == 1 )
+            TickType_t xLastActTime;                      /**< The last time when hang-protection was done.*/
+        #endif /* ipconfigTCP_HANG_PROTECTION */
+        size_t uxLittleSpace;                             /**< The value deemed as low amount of space. */
+        size_t uxEnoughSpace;                             /**< The value deemed as enough space. */
+        size_t uxRxStreamSize;                            /**< The Receive stream size */
+        size_t uxTxStreamSize;                            /**< The transmit stream size */
+        StreamBuffer_t * rxStream;                        /**< The pointer to the receive stream buffer. */
+        StreamBuffer_t * txStream;                        /**< The pointer to the transmit stream buffer. */
+        #if ( ipconfigUSE_TCP_WIN == 1 )
+            NetworkBufferDescriptor_t * pxAckMessage;     /**< The pointer to the ACK message */
+        #endif /* ipconfigUSE_TCP_WIN */
+        LastTCPPacket_t xPacket;                          /**< Buffer space to store the last TCP header received. */
+        uint8_t tcpflags;                                 /**< TCP flags */
+        #if ( ipconfigUSE_TCP_WIN != 0 )
+            uint8_t ucMyWinScaleFactor;                   /**< Scaling factor of this device. */
+            uint8_t ucPeerWinScaleFactor;                 /**< Scaling factor of the peer. */
+        #endif
+        #if ( ipconfigUSE_CALLBACKS == 1 )
+            FOnTCPReceive_t pxHandleReceive;      /**<
                                                    * In case of a TCP socket:
                                                    * typedef void (* FOnTCPReceive_t) (Socket_t xSocket, void *pData, size_t uxLength );
                                                    */
-                FOnTCPSent_t pxHandleSent;        /**< Function pointer to handle a successful send event.  */
-                FOnConnected_t pxHandleConnected; /**< Actually type: typedef void (* FOnConnected_t) (Socket_t xSocket, BaseType_t ulConnected ); */
-            #endif /* ipconfigUSE_CALLBACKS */
-            uint32_t ulWindowSize;                /**< Current Window size advertised by peer */
-            size_t uxRxWinSize;                   /**< Fixed value: size of the TCP reception window */
-            size_t uxTxWinSize;                   /**< Fixed value: size of the TCP transmit window */
+            FOnTCPSent_t pxHandleSent;            /**< Function pointer to handle a successful send event.  */
+            FOnConnected_t pxHandleConnected;     /**< Actually type: typedef void (* FOnConnected_t) (Socket_t xSocket, BaseType_t ulConnected ); */
+        #endif /* ipconfigUSE_CALLBACKS */
+        uint32_t ulWindowSize;                    /**< Current Window size advertised by peer */
+        size_t uxRxWinSize;                       /**< Fixed value: size of the TCP reception window */
+        size_t uxTxWinSize;                       /**< Fixed value: size of the TCP transmit window */
 
-            TCPWindow_t xTCPWindow;               /**< The TCP window struct*/
-        } IPTCPSocket_t;
+        TCPWindow_t xTCPWindow;                   /**< The TCP window struct*/
+    } IPTCPSocket_t;
 
-    #endif /* ipconfigUSE_TCP */
+#endif /* ipconfigUSE_TCP */
 
 /**
  * Structure to hold the information about a UDP socket.
  */
-    typedef struct UDPSOCKET
-    {
-        List_t xWaitingPacketsList;   /**< Incoming packets */
-        #if ( ipconfigUDP_MAX_RX_PACKETS > 0 )
-            UBaseType_t uxMaxPackets; /**< Protection: limits the number of packets buffered per socket */
-        #endif /* ipconfigUDP_MAX_RX_PACKETS */
-        #if ( ipconfigUSE_CALLBACKS == 1 )
-            FOnUDPReceive_t pxHandleReceive; /*
+typedef struct UDPSOCKET
+{
+    List_t xWaitingPacketsList;       /**< Incoming packets */
+    #if ( ipconfigUDP_MAX_RX_PACKETS > 0 )
+        UBaseType_t uxMaxPackets;     /**< Protection: limits the number of packets buffered per socket */
+    #endif /* ipconfigUDP_MAX_RX_PACKETS */
+    #if ( ipconfigUSE_CALLBACKS == 1 )
+        FOnUDPReceive_t pxHandleReceive;     /*
                                               * In case of a UDP socket:
                                               * typedef void ( * FOnUDPReceive_t ) ( Socket_t xSocket, void *pData, size_t uxLength, struct freertos_sockaddr *pxAddr );
                                               */
-            FOnUDPSent_t pxHandleSent;
-        #endif /* ipconfigUSE_CALLBACKS */
-    } IPUDPSocket_t;
+        FOnUDPSent_t pxHandleSent;
+    #endif /* ipconfigUSE_CALLBACKS */
+} IPUDPSocket_t;
 
 /* Formally typedef'd as eSocketEvent_t. */
-    enum eSOCKET_EVENT
-    {
-        eSOCKET_RECEIVE = 0x0001,
-        eSOCKET_SEND = 0x0002,
-        eSOCKET_ACCEPT = 0x0004,
-        eSOCKET_CONNECT = 0x0008,
-        eSOCKET_BOUND = 0x0010,
-        eSOCKET_CLOSED = 0x0020,
-        eSOCKET_INTR = 0x0040,
-        eSOCKET_ALL = 0x007F
-    };
+enum eSOCKET_EVENT
+{
+    eSOCKET_RECEIVE = 0x0001,
+    eSOCKET_SEND = 0x0002,
+    eSOCKET_ACCEPT = 0x0004,
+    eSOCKET_CONNECT = 0x0008,
+    eSOCKET_BOUND = 0x0010,
+    eSOCKET_CLOSED = 0x0020,
+    eSOCKET_INTR = 0x0040,
+    eSOCKET_ALL = 0x007F
+};
 
 
 /**
  * Structure to hold information for a socket.
  */
-    typedef struct xSOCKET
+typedef struct xSOCKET
+{
+    EventBits_t xEventBits;             /**< The eventbits to keep track of events. */
+    EventGroupHandle_t xEventGroup;     /**< The event group for this socket. */
+
+    /* Most compilers do like bit-flags */
+    struct
     {
-        EventBits_t xEventBits;         /**< The eventbits to keep track of events. */
-        EventGroupHandle_t xEventGroup; /**< The event group for this socket. */
-
-        /* Most compilers do like bit-flags */
-        struct
-        {
-            uint32_t
-            #if ( ipconfigUSE_IPv6 != 0 )
-                bIsIPv6 : 1, /**< Non-zero in case the connection is using IPv6. */
-            #endif /* ipconfigUSE_IPv6 */
-            bSomeFlag : 1;
-        } bits;                          /**< A collection of boolean properties. */
-
-        ListItem_t xBoundSocketListItem; /**< Used to reference the socket from a bound sockets list. */
-        TickType_t xReceiveBlockTime;    /**< if recv[to] is called while no data is available, wait this amount of time. Unit in clock-ticks */
-        TickType_t xSendBlockTime;       /**< if send[to] is called while there is not enough space to send, wait this amount of time. Unit in clock-ticks */
-
-        uint32_t ulLocalAddress;         /**< Local IP address */
+        uint32_t
         #if ( ipconfigUSE_IPv6 != 0 )
-            IPv6_Address_t xLocalAddress_IPv6;
-        #endif
-        uint16_t usLocalPort;                  /**< Local port on this machine */
-        uint8_t ucSocketOptions;               /**< Socket options */
-        uint8_t ucProtocol;                    /**< choice of FREERTOS_IPPROTO_UDP/TCP */
-        #if ( ipconfigSOCKET_HAS_USER_SEMAPHORE == 1 )
-            SemaphoreHandle_t pxUserSemaphore; /**< The user semaphore */
-        #endif /* ipconfigSOCKET_HAS_USER_SEMAPHORE */
-        #if ( ipconfigSOCKET_HAS_USER_WAKE_CALLBACK == 1 )
-            SocketWakeupCallback_t pxUserWakeCallback; /**< Pointer to the callback function. */
-        #endif /* ipconfigSOCKET_HAS_USER_WAKE_CALLBACK */
+            bIsIPv6 : 1,     /**< Non-zero in case the connection is using IPv6. */
+        #endif /* ipconfigUSE_IPv6 */
+        bSomeFlag : 1;
+    } bits;                              /**< A collection of boolean properties. */
 
-        #if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 )
-            struct xSOCKET_SET * pxSocketSet; /**< Pointer to the socket set structure */
-            EventBits_t xSelectBits;          /**< User may indicate which bits are interesting for this socket. */
+    ListItem_t xBoundSocketListItem;     /**< Used to reference the socket from a bound sockets list. */
+    TickType_t xReceiveBlockTime;        /**< if recv[to] is called while no data is available, wait this amount of time. Unit in clock-ticks */
+    TickType_t xSendBlockTime;           /**< if send[to] is called while there is not enough space to send, wait this amount of time. Unit in clock-ticks */
 
-            EventBits_t xSocketBits;          /**< These bits indicate the events which have actually occurred.
+    uint32_t ulLocalAddress;             /**< Local IP address */
+    #if ( ipconfigUSE_IPv6 != 0 )
+        IPv6_Address_t xLocalAddress_IPv6;
+    #endif
+    uint16_t usLocalPort;                      /**< Local port on this machine */
+    uint8_t ucSocketOptions;                   /**< Socket options */
+    uint8_t ucProtocol;                        /**< choice of FREERTOS_IPPROTO_UDP/TCP */
+    #if ( ipconfigSOCKET_HAS_USER_SEMAPHORE == 1 )
+        SemaphoreHandle_t pxUserSemaphore;     /**< The user semaphore */
+    #endif /* ipconfigSOCKET_HAS_USER_SEMAPHORE */
+    #if ( ipconfigSOCKET_HAS_USER_WAKE_CALLBACK == 1 )
+        SocketWakeupCallback_t pxUserWakeCallback;     /**< Pointer to the callback function. */
+    #endif /* ipconfigSOCKET_HAS_USER_WAKE_CALLBACK */
+
+    #if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 )
+        struct xSOCKET_SET * pxSocketSet;     /**< Pointer to the socket set structure */
+        EventBits_t xSelectBits;              /**< User may indicate which bits are interesting for this socket. */
+
+        EventBits_t xSocketBits;              /**< These bits indicate the events which have actually occurred.
                                                * They are maintained by the IP-task */
-        #endif /* ipconfigSUPPORT_SELECT_FUNCTION */
-        struct xNetworkEndPoint * pxEndPoint; /**< The end-point to which the socket is bound. */
-        /* TCP/UDP specific fields: */
-        /* Before accessing any member of this structure, it should be confirmed */
-        /* that the protocol corresponds with the type of structure */
+    #endif /* ipconfigSUPPORT_SELECT_FUNCTION */
+    struct xNetworkEndPoint * pxEndPoint;     /**< The end-point to which the socket is bound. */
+    /* TCP/UDP specific fields: */
+    /* Before accessing any member of this structure, it should be confirmed */
+    /* that the protocol corresponds with the type of structure */
 
-        union
-        {
-            IPUDPSocket_t xUDP;           /**< Union member: UDP socket*/
-            #if ( ipconfigUSE_TCP == 1 )
-                IPTCPSocket_t xTCP;       /**< Union member: TCP socket */
+    union
+    {
+        IPUDPSocket_t xUDP;               /**< Union member: UDP socket*/
+        #if ( ipconfigUSE_TCP == 1 )
+            IPTCPSocket_t xTCP;           /**< Union member: TCP socket */
 
-                uint64_t ullTCPAlignment; /**< Make sure that xTCP is 8-bytes aligned by
+            uint64_t ullTCPAlignment;     /**< Make sure that xTCP is 8-bytes aligned by
                                            * declaring a 64-bit variable in the same union */
-            #endif /* ipconfigUSE_TCP */
-        } u;                              /**< Union of TCP/UDP socket */
-    } FreeRTOS_Socket_t;
+        #endif /* ipconfigUSE_TCP */
+    } u;                                  /**< Union of TCP/UDP socket */
+} FreeRTOS_Socket_t;
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( FreeRTOS_Socket_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( FreeRTOS_Socket_t );
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( FreeRTOS_Socket_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( FreeRTOS_Socket_t );
 
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( NetworkBufferDescriptor_t );
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( NetworkBufferDescriptor_t );
 
 /* Allow casting from a different pointer. */
-    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( sockaddr4_t );
-    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( sockaddr4_t );
+extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( sockaddr4_t );
+extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( sockaddr4_t );
 
-    #if ( ipconfigUSE_IPv6 != 0 )
-        /* Allow casting from a different pointer, mostly sockaddr4_t. */
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( sockaddr6_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( sockaddr6_t );
-    #endif
+#if ( ipconfigUSE_IPv6 != 0 )
+    /* Allow casting from a different pointer, mostly sockaddr4_t. */
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( sockaddr6_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( sockaddr6_t );
+#endif
 
-    #if ( ipconfigUSE_TCP == 1 )
+#if ( ipconfigUSE_TCP == 1 )
 
 /*
  * Close the socket another time.
  */
-        void vSocketCloseNextTime( FreeRTOS_Socket_t * pxSocket );
+    void vSocketCloseNextTime( FreeRTOS_Socket_t * pxSocket );
 
 /*
  * Lookup a TCP socket, using a multiple matching: both port numbers and
  * return IP address.
  */
-        FreeRTOS_Socket_t * pxTCPSocketLookup( UBaseType_t uxLocalPort,
-                                               uint32_t ulRemoteIP,
-                                               UBaseType_t uxRemotePort
-        #if ( ipconfigUSE_IPv6 != 0 )
-                                                   ,
-                                                   IPv6_Address_t * pxAddress_IPv6
-        #endif /* ipconfigUSE_IPv6 */
-                                               );
+    FreeRTOS_Socket_t * pxTCPSocketLookup( UBaseType_t uxLocalPort,
+                                           uint32_t ulRemoteIP,
+                                           UBaseType_t uxRemotePort
+    #if ( ipconfigUSE_IPv6 != 0 )
+                                               ,
+                                               IPv6_Address_t * pxAddress_IPv6
+    #endif /* ipconfigUSE_IPv6 */
+                                           );
 
-    #endif /* ipconfigUSE_TCP */
+#endif /* ipconfigUSE_TCP */
 
 
 /*
  * Look up a local socket by finding a match with the local port.
  */
-    FreeRTOS_Socket_t * pxUDPSocketLookup( UBaseType_t uxLocalPort );
+FreeRTOS_Socket_t * pxUDPSocketLookup( UBaseType_t uxLocalPort );
 
 /*
  * Called when the application has generated a UDP packet to send.
  */
-    void vProcessGeneratedUDPPacket( NetworkBufferDescriptor_t * const pxNetworkBuffer );
+void vProcessGeneratedUDPPacket( NetworkBufferDescriptor_t * const pxNetworkBuffer );
 
 /*
  * Calculate the upper-layer checksum
@@ -1013,16 +1016,16 @@
  * bOut = false: checksum will be calculated for incoming packets
  *     returning 0xffff means: checksum was correct
  */
-    uint16_t usGenerateProtocolChecksum( uint8_t * pucEthernetBuffer,
-                                         size_t uxBufferLength,
-                                         BaseType_t xOutgoingPacket );
+uint16_t usGenerateProtocolChecksum( uint8_t * pucEthernetBuffer,
+                                     size_t uxBufferLength,
+                                     BaseType_t xOutgoingPacket );
 
 /*
  * An Ethernet frame has been updated (maybe it was an ARP request or a PING
  * request?) and is to be sent back to its source.
  */
-    void vReturnEthernetFrame( NetworkBufferDescriptor_t * pxNetworkBuffer,
-                               BaseType_t xReleaseAfterSend );
+void vReturnEthernetFrame( NetworkBufferDescriptor_t * pxNetworkBuffer,
+                           BaseType_t xReleaseAfterSend );
 
 /*
  * The internal version of bind()
@@ -1030,10 +1033,10 @@
  * The TCP driver needs to bind a socket at the moment a listening socket
  * creates a new connected socket
  */
-    BaseType_t vSocketBind( FreeRTOS_Socket_t * pxSocket,
-                            struct freertos_sockaddr * pxBindAddress,
-                            size_t uxAddressLength,
-                            BaseType_t xInternal );
+BaseType_t vSocketBind( FreeRTOS_Socket_t * pxSocket,
+                        struct freertos_sockaddr * pxBindAddress,
+                        size_t uxAddressLength,
+                        BaseType_t xInternal );
 
 /*
  * Internal function to add streaming data to a TCP socket. If ulIn == true,
@@ -1042,150 +1045,142 @@
  * packet come in out-of-order, an offset will be used to put it in front and
  * the head will not change yet.
  */
-    int32_t lTCPAddRxdata( FreeRTOS_Socket_t * pxSocket,
-                           size_t uxOffset,
-                           const uint8_t * pcData,
-                           uint32_t ulByteCount );
+int32_t lTCPAddRxdata( FreeRTOS_Socket_t * pxSocket,
+                       size_t uxOffset,
+                       const uint8_t * pcData,
+                       uint32_t ulByteCount );
 
 /*
  * Currently called for any important event.
  */
-    void vSocketWakeUpUser( FreeRTOS_Socket_t * pxSocket );
+void vSocketWakeUpUser( FreeRTOS_Socket_t * pxSocket );
 
 /*
  * Some helping function, their meaning should be clear
  */
-    uint32_t ulChar2u32( const uint8_t * pucPtr );
+uint32_t ulChar2u32( const uint8_t * pucPtr );
 
-    uint16_t usChar2u16( const uint8_t * pucPtr );
+uint16_t usChar2u16( const uint8_t * pucPtr );
 
 /* Check a single socket for retransmissions and timeouts */
-    BaseType_t xTCPSocketCheck( FreeRTOS_Socket_t * pxSocket );
+BaseType_t xTCPSocketCheck( FreeRTOS_Socket_t * pxSocket );
 
-    BaseType_t xTCPCheckNewClient( FreeRTOS_Socket_t * pxSocket );
+BaseType_t xTCPCheckNewClient( FreeRTOS_Socket_t * pxSocket );
 
 /* Defined in FreeRTOS_Sockets.c
  * Close a socket
  */
-    void * vSocketClose( FreeRTOS_Socket_t * pxSocket );
+void * vSocketClose( FreeRTOS_Socket_t * pxSocket );
 
 /*
  * Send the event eEvent to the IP task event queue, using a block time of
  * zero.  Return pdPASS if the message was sent successfully, otherwise return
  * pdFALSE.
  */
-    BaseType_t xSendEventToIPTask( eIPEvent_t eEvent );
+BaseType_t xSendEventToIPTask( eIPEvent_t eEvent );
 
 /*
  * The same as above, but a struct as a parameter, containing:
  *      eIPEvent_t eEventType;
  *		void *pvData;
  */
-    BaseType_t xSendEventStructToIPTask( const IPStackEvent_t * pxEvent,
-                                         TickType_t uxTimeout );
+BaseType_t xSendEventStructToIPTask( const IPStackEvent_t * pxEvent,
+                                     TickType_t uxTimeout );
 
 /*
  * Returns a pointer to the original NetworkBuffer from a pointer to a UDP
  * payload buffer.
  */
-    NetworkBufferDescriptor_t * pxUDPPayloadBuffer_to_NetworkBuffer( const void * pvBuffer );
+NetworkBufferDescriptor_t * pxUDPPayloadBuffer_to_NetworkBuffer( const void * pvBuffer );
 
 /* Get the size of the IP-header.
  * 'usFrameType' must be filled in if IPv6is to be recognised. */
-    size_t uxIPHeaderSizePacket( const NetworkBufferDescriptor_t * pxNetworkBuffer );
+size_t uxIPHeaderSizePacket( const NetworkBufferDescriptor_t * pxNetworkBuffer );
 /*-----------------------------------------------------------*/
 
 /* Get the size of the IP-header.
  * The socket is checked for its type: IPv4 or IPv6. */
-    size_t uxIPHeaderSizeSocket( const FreeRTOS_Socket_t * pxSocket );
+size_t uxIPHeaderSizeSocket( const FreeRTOS_Socket_t * pxSocket );
 /*-----------------------------------------------------------*/
 
-    #if ( ipconfigZERO_COPY_TX_DRIVER != 0 )
-
-/*
- * For the case where the network driver passes a buffer directly to a DMA
- * descriptor, this function can be used to translate a 'network buffer' to
- * a 'network buffer descriptor'.
- */
-        NetworkBufferDescriptor_t * pxPacketBuffer_to_NetworkBuffer( const void * pvBuffer );
-    #endif
-
-    #if ( ( ipconfigHAS_DEBUG_PRINTF != 0 ) || ( ipconfigHAS_PRINTF != 0 ) )
-        /* prepare a string which describes a socket, just for logging. */
-        const char * prvSocketProps( FreeRTOS_Socket_t * pxSocket );
-    #endif /* ipconfigHAS_DEBUG_PRINTF || ipconfigHAS_PRINTF */
+#if ( ( ipconfigHAS_DEBUG_PRINTF != 0 ) || ( ipconfigHAS_PRINTF != 0 ) )
+    /* prepare a string which describes a socket, just for logging. */
+    const char * prvSocketProps( FreeRTOS_Socket_t * pxSocket );
+#endif /* ipconfigHAS_DEBUG_PRINTF || ipconfigHAS_PRINTF */
 
 /*
  * Internal: Sets a new state for a TCP socket and performs the necessary
  * actions like calling a OnConnected handler to notify the socket owner.
  */
-    #if ( ipconfigUSE_TCP == 1 )
-        void vTCPStateChange( FreeRTOS_Socket_t * pxSocket,
-                              enum eTCP_STATE eTCPState );
-    #endif /* ipconfigUSE_TCP */
+#if ( ipconfigUSE_TCP == 1 )
+    void vTCPStateChange( FreeRTOS_Socket_t * pxSocket,
+                          enum eTCP_STATE eTCPState );
+#endif /* ipconfigUSE_TCP */
 
 /* Returns pdTRUE is this function is called from the IP-task */
-    BaseType_t xIsCallingFromIPTask( void );
+BaseType_t xIsCallingFromIPTask( void );
 
-    #if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 )
+#if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 )
 
 /** @brief Structure for event groups of the Socket Select functions */
-        typedef struct xSOCKET_SET
-        {
-            /** @brief Event group for the socket select function.
-             */
-            EventGroupHandle_t xSelectGroup;
-        } SocketSelect_t;
+    typedef struct xSOCKET_SET
+    {
+        /** @brief Event group for the socket select function.
+         */
+        EventGroupHandle_t xSelectGroup;
+    } SocketSelect_t;
 
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( SocketSelect_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( SocketSelect_t );
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( SocketSelect_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( SocketSelect_t );
 
-        extern void vSocketSelect( SocketSelect_t * pxSocketSet );
+    extern void vSocketSelect( SocketSelect_t * pxSocketSet );
 
 /** @brief Define the data that must be passed for a 'eSocketSelectEvent'. */
-        typedef struct xSocketSelectMessage
-        {
-            TaskHandle_t xTaskhandle;     /**< Task handle for use in the socket select functionality. */
-            SocketSelect_t * pxSocketSet; /**< The event group for the socket select functionality. */
-        } SocketSelectMessage_t;
+    typedef struct xSocketSelectMessage
+    {
+        TaskHandle_t xTaskhandle;         /**< Task handle for use in the socket select functionality. */
+        SocketSelect_t * pxSocketSet;     /**< The event group for the socket select functionality. */
+    } SocketSelectMessage_t;
 
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( SocketSelectMessage_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( SocketSelectMessage_t );
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( SocketSelectMessage_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( SocketSelectMessage_t );
 
-    #endif /* ipconfigSUPPORT_SELECT_FUNCTION */
+#endif /* ipconfigSUPPORT_SELECT_FUNCTION */
 
-    #if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 ) || ( ipconfigUSE_TCP == 1 ) || ( ipconfigDNS_USE_CALLBACKS == 1 )
-        extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ListItem_t );
-        extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ListItem_t );
-    #endif
+#if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 ) || ( ipconfigUSE_TCP == 1 ) || ( ipconfigDNS_USE_CALLBACKS == 1 )
+    extern ipDECL_CAST_PTR_FUNC_FOR_TYPE( ListItem_t );
+    extern ipDECL_CAST_CONST_PTR_FUNC_FOR_TYPE( ListItem_t );
+#endif
 
-    #if ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 )
-        void vIPSetDHCP_RATimerEnableState( struct xNetworkEndPoint * pxEndPoint,
-                                            BaseType_t xEnableState );
-    #endif /* ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 ) */
+#if ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 )
+    void vIPSetDHCP_RATimerEnableState( struct xNetworkEndPoint * pxEndPoint,
+                                        BaseType_t xEnableState );
+#endif /* ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 ) */
 
-    #if ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 )
-        void vIPReloadDHCP_RATimer( struct xNetworkEndPoint * pxEndPoint,
-                                    TickType_t uxClockTicks );
-    #endif /* ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 ) */
+#if ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 )
+    void vIPReloadDHCP_RATimer( struct xNetworkEndPoint * pxEndPoint,
+                                TickType_t uxClockTicks );
+#endif /* ( ipconfigUSE_DHCP == 1 ) || ( ipconfigUSE_RA == 1 ) */
 
-    #if ( ipconfigDNS_USE_CALLBACKS != 0 )
-        void vIPReloadDNSTimer( uint32_t ulCheckTime );
-        void vIPSetDnsTimerEnableState( BaseType_t xEnableState );
-    #endif
+#if ( ipconfigDNS_USE_CALLBACKS != 0 )
+    void vIPReloadDNSTimer( uint32_t ulCheckTime );
+    void vIPSetDnsTimerEnableState( BaseType_t xEnableState );
+#endif
 
 /* Send the network-up event and start the ARP timer. */
-    void vIPNetworkUpCalls( struct xNetworkEndPoint * pxEndPoint );
+void vIPNetworkUpCalls( struct xNetworkEndPoint * pxEndPoint );
 
-    #if ( ipconfigUSE_IPv6 != 0 )
+#if ( ipconfigUSE_IPv6 != 0 )
 
 /* prvProcessICMPMessage_IPv6() is declared in FreeRTOS_routing.c
  * It handles all ICMP messages except the PING requests. */
-        eFrameProcessingResult_t prvProcessICMPMessage_IPv6( NetworkBufferDescriptor_t * const pxNetworkBuffer );
-    #endif
+    eFrameProcessingResult_t prvProcessICMPMessage_IPv6( NetworkBufferDescriptor_t * const pxNetworkBuffer );
+#endif
 
-    #ifdef __cplusplus
-        } /* extern "C" */
-    #endif
+/* *INDENT-OFF* */
+#ifdef __cplusplus
+    } /* extern "C" */
+#endif
+/* *INDENT-ON* */
 
 #endif /* FREERTOS_IP_PRIVATE_H */
