+++ test/cbmc/patches/FreeRTOSIPConfig.h
@@ -37,7 +37,9 @@
 /* Set to 1 to print out debug messages.  If ipconfigHAS_DEBUG_PRINTF is set to
  * 1 then FreeRTOS_debug_printf should be defined to the function used to print
  * out the debugging messages. */
-#define ipconfigHAS_DEBUG_PRINTF    0
+#ifndef ipconfigHAS_DEBUG_PRINTF
+    #define ipconfigHAS_DEBUG_PRINTF    0
+#endif
 #if ( ipconfigHAS_DEBUG_PRINTF == 1 )
     #define FreeRTOS_debug_printf( X )    configPRINTF( X )
 #endif
@@ -230,7 +232,9 @@ extern uint32_t ulRand();
 #endif
 
 /* USE_WIN: Let TCP use windowing mechanism. */
-#define ipconfigUSE_TCP_WIN                ( 1 )
+#ifndef ipconfigUSE_TCP_WIN
+    #define ipconfigUSE_TCP_WIN    ( 1 )
+#endif
 
 /* The MTU is the maximum number of bytes the payload of a network frame can
  * contain.  For normal Ethernet V2 frames the maximum MTU is 1500.  Setting a
@@ -255,7 +259,9 @@ extern uint32_t ulRand();
 
 /* If ipconfigSUPPORT_SELECT_FUNCTION is set to 1 then the FreeRTOS_select()
  * (and associated) API function is available. */
-#define ipconfigSUPPORT_SELECT_FUNCTION                0
+#ifndef ipconfigSUPPORT_SELECT_FUNCTION
+    #define ipconfigSUPPORT_SELECT_FUNCTION    0
+#endif
 
 /* If ipconfigFILTER_OUT_NON_ETHERNET_II_FRAMES is set to 1 then Ethernet frames
  * that are not in Ethernet II format will be dropped.  This option is included for
